adding testing files & correcting the code according to the review
