+++ buddha-system/package.json
@@ -11,6 +11,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "testdb": "NODE_ENV=test node tests/db.tests.js",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
