+++ buddha-system/public/favicon.ico
