--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "buddha",
-  "version": "1.0.0",
-  "description": "hp-oranBit Project",
-  "main": "index.js",
-  "dependencies": {
-    "dotenv": "^8.0.0",
-    "pg": "^7.11.0",
-    "url": "^0.11.0"
-  },
-  "devDependencies": {
-    "tap-spec": "^5.0.0",
-    "tape": "^4.11.0"
-  },
-  "scripts": {
-    "testdb": "NODE_ENV=test node tests/db.tests.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lotus-1/Buddha.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/lotus-1/Buddha/issues"
-  },
-  "homepage": "https://github.com/lotus-1/Buddha#readme"
-}
